requirements.txt
matrix-nio>=0.24.0
python-dotenv>=1.0.0
pyyaml>=6.0
aiofiles>=23.0
